fix: import and linting errors when running as package
