+++ .vscode/settings.json
@@ -1,5 +1,1 @@
-{
-    "python.analysis.extraPaths": [
-        "./diff_tool/diff_tool"
-    ]
-}
+{}
